Async filters
We have our own subtype of response/request filter interfaces, which take
our own subtype of response/request context, which have extra methods.
I've made it so all the suspend/resume action is on the request, instead of
both request and response, simply because the response filter have the request
context anyway, so it didn't make much sense to duplicate those actions in both
contexts, since they do the same damn thing.
Introduced a public face for QuarkusRestRequestContext, which doesn't have a very
nice name, but at least can get injection and is more decent for users.
Moved Completion/ConnectionCallbacks to the context, because I've always hated
that I needed the AsyncResponse to get them, most times you want them regardless
of whether the request is sync or async.